requirements.txt
fastapi
uvicorn
sqlmodel
python-dotenv
pyjwt[crypto]